01_test_prin_tf/spagri/config/common/requirements.txt
ruamel.yaml==0.18.10
pydantic==2.11.1
typeguard==4.4.2
